# backend/swagger/spec.json
{"openapi":"3.0.0","info":{"title":"API SmartClass","version":"1.0.0"},"paths":{"/class":{"get":{"summary":"Returns the class in which the user is a teacher or a student.","tags":["class"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"header","name":"Authorization","schema":{"type":"string","format":"JWT"},"required":true}],"responses":{"200":{"description":"A JSON Object of a Class","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Class"}}}},"401":{"description":"Authorization information is missing or invalid."},"403":{"description":"The role of the user does not permit that action"},"404":{"description":"Class of the user not found"},"5XX":{"description":"Unexpected error."}}}},"/class/teacher":{"get":{"summary":"Returns the class teacher","tags":["class","teacher"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"header","name":"Authorization","schema":{"type":"string","format":"JWT"},"required":true}],"responses":{"200":{"description":"A JSON Object of a Class","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClassTeacher"}}}},"401":{"description":"Authorization information is missing or invalid."},"403":{"description":"The role of the user does not permit that action"},"404":{"description":"Class of the user not found"},"5XX":{"description":"Unexpected error."}}}},"/class/pupils":{"get":{"summary":"Returns the pupils in the class","tags":["class","pupils"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"header","name":"Authorization","schema":{"type":"string","format":"JWT"},"required":true}],"responses":{"200":{"description":"A JSON Object of a Class","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClassPupils"}}}},"401":{"description":"Authorization information is missing or invalid."},"403":{"description":"The role of the user does not permit that action"},"404":{"description":"Class of the user not found"},"5XX":{"description":"Unexpected error."}}}}},"components":{"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}},"responses":{"401":{"description":"JWT is"},"MissingJWT":{"description":"le JWT n'est pas présent"}},"schemas":{"Class":{"type":"object","properties":{"id":{"type":"integer","format":"INTEGER"},"classname":{"type":"string","format":"CHAR(255)"}},"example":{"id":42,"classname":"2B"}},"ClassPupils":{"type":"array","items":{"type":"object","properties":{"id":{"type":"integer","format":"INTEGER"},"firstname":{"type":"string","format":"CHAR(255)"},"lastname":{"type":"string","format":"CHAR(255)"}}},"example":[{"id":420,"firstname":"Thomas","lastname":"Martin"},{"id":69,"firstname":"Bob","lastname":"Dubuisson"}]},"ClassTeacher":{"type":"object","properties":{"id":{"type":"integer","format":"INTEGER"},"login":{"type":"string","format":"CHAR(255)"},"firstname":{"type":"string","format":"CHAR(255)"},"lastname":{"type":"string","format":"CHAR(255)"}},"example":{"id":69,"firstname":"Antoine","lastname":"Lambert"}}}},"tags":[]}